Add new upload fields to form
