+++ app/forms/submissions/trial/step02_content_form.rb
@@ -1,96 +1,25 @@
 module Submissions
   module Trial
     class Step02ContentForm < PlantTrialForm
-      property :trait_descriptor_list
+      property :lines_or_varieties
+      collection :technical_replicate_numbers
+      collection :design_factor_names
 
-      collection :new_trait_descriptors do
-        property :trait
-        property :plant_part_id
-        property :units_of_measurements
-
-        property :scoring_method
-        property :materials
-
-        property :data_owned_by
-        property :data_provenance
-        property :comments
-
-        validates :trait, presence: true
-        validates :units_of_measurements, presence: true
-        validates :scoring_method, presence: true
-      end
-
-      validates :trait_descriptor_list, length: { minimum: 1 }
-
-      # NOTE Left here for the moment we allow users to define their own trait names.
-      validate do
-        new_trait_descriptors.each do |new_trait_descriptor|
-          if trait_descriptor_exists?("traits.name ILIKE ?", new_trait_descriptor.trait)
-            errors.add(:new_trait_descriptors, :taken, name: new_trait_descriptor.trait)
-          end
-        end
-      end
+      validates :lines_or_varieties, inclusion: { in: %w(plant_lines plant_varieties) }
 
       validate do
-        duplicated_trait_descriptors =
-          Hash[trait_descriptor_list.map { |td| [td, trait_descriptor_list.count(td)] }].
-          select { |td, count| count > 1 }.
-          keys
-
-        duplicated_trait_descriptors.each do |trait_descriptor|
-          errors.add(:trait_descriptor_list, :duplicated, name: trait_descriptor)
+        all_valid = technical_replicate_numbers.all? do |number|
+          number.to_i.to_s == number && number.to_i > 0
         end
+        errors.add(:technical_replicate_numbers, :invalid) unless all_valid
       end
 
-      # Ensure all items in :trait_descriptor_list either exist or have
-      # valid entries in :new_trait_descriptors
-      validate do
-        trait_descriptor_list.each do |id_or_name|
-          unless trait_descriptor_exists?(id: id_or_name) || new_trait_descriptors.map(&:trait).include?(id_or_name)
-            errors.add(:trait_descriptor_list, :blank, name: id_or_name)
-          end
-        end
+      def technical_replicate_numbers
+        super || []
       end
 
-      def trait_descriptor_exists?(*attrs)
-        TraitDescriptor.includes(:trait).where(*attrs).exists?
-      end
-
-      def existing_trait_descriptors
-        ids = trait_descriptor_list.try(:map, &:to_i)
-        TraitDescriptor.where(id: ids)
-      end
-
-      def self.permitted_properties
-        [
-          {
-            :trait_descriptor_list => [],
-            :new_trait_descriptors => new_trait_descriptor_properties
-          }
-        ]
-      end
-
-      def self.new_trait_descriptor_properties
-        [
-          :trait,
-          :plant_part_id,
-          :units_of_measurements,
-          :scoring_method,
-          :materials,
-          :data_owned_by,
-          :data_provenance,
-          :comments
-        ]
-      end
-
-      def trait_descriptor_list
-        super.try { |tdl| tdl.select(&:present?) }
-      end
-
-      def new_trait_descriptors
-        (super || []).select do |trait_descriptor|
-          trait_descriptor_list.include?(trait_descriptor.trait)
-        end
+      def design_factor_names
+        super || []
       end
     end
   end
