+++ app/forms/submissions/trial/step05_content_form.rb
@@ -1,10 +1,11 @@
 module Submissions
   module Trial
     class Step05ContentForm < PlantTrialForm
-      property :layout_upload_id
+      property :environment_upload_id
 
-      def layout_upload
-        Submission::Upload.plant_trial_layout.find_by(id: layout_upload_id)
+      def environment_upload
+        upload = Submission::Upload.plant_trial_environment.find_by(id: environment_upload_id)
+        SubmissionPlantTrialEnvironmentUploadDecorator.decorate(upload) if upload
       end
     end
   end
