+++ app/models/plant_trial/environment.rb
@@ -10,20 +10,20 @@ class PlantTrial::Environment < ActiveRecord::Base
       light_intensity: { non_negative: true },
       light_intensity_range: { non_negative: true },
       outside_light_loss: { non_negative: true },
-      rfr_ratio: { ratio: true },
+      rfr_ratio: { percentage: true },
       daily_uvb: { non_negative: true },
       total_light: { non_negative: true },
       co2_light: { non_negative: true },
       co2_dark: { non_negative: true },
-      relative_humidity_light: { ratio: true },
-      relative_humidity_dark: { ratio: true },
+      relative_humidity_light: { percentage: true },
+      relative_humidity_dark: { percentage: true },
       rooting_container_volume: { non_negative: true },
       rooting_container_height: { non_negative: true },
       rooting_count: { non_negative: { only_integer: true } },
       sowing_density: { non_negative: true },
-      soil_porosity: { ratio: true },
+      soil_porosity: { percentage: true },
       soil_penetration: { non_negative: true },
-      soil_organic_matter: { ratio: true },
+      soil_organic_matter: { percentage: true },
       medium_temperature: { numericality: true },
       water_retention: { numericality: true },
       nitrogen_concentration_start: { non_negative: true },
