+++ app/models/submission/upload.rb
@@ -17,6 +17,8 @@ class Submission::Upload < ActiveRecord::Base
     if: -> { plant_trial_layout? }
   validates :file, attachment_content_type: { content_type: CSV_CONTENT_TYPES },
     if: -> { trait_scores? || plant_lines? }
+  validates :file, attachment_content_type: { content_type: XLS_CONTENT_TYPES },
+    if: -> { plant_trial_environment? || plant_trial_treatment? }
 
   # NOTE, WARNING: @logs will contain user-provided data; do NOT interpret it as html
   attr_reader :logs
