+++ app/validators/percentage_validator.rb
@@ -1,6 +1,6 @@
-class RatioValidator < ActiveModel::Validations::NumericalityValidator
+class PercentageValidator < ActiveModel::Validations::NumericalityValidator
   def initialize(options = {})
-    options.merge!(greater_than_or_equal_to: 0, less_than_or_equal_to: 1)
+    options.merge!(greater_than_or_equal_to: 0, less_than_or_equal_to: 100)
 
     super(options)
   end
