+++ app/views/submissions/steps/trial/_step01.html.haml
@@ -1,32 +1,8 @@
-.form-group
-  = content.text_field :plant_trial_name, required: true,
-    help: 'Please, provide convenient <b>name</b> for this scoring trial. It has to be <b>unique</b>.'
-
-.form-group
-  - project_descriptors = PlantTrial.visible(current_user.id).attribute_values(:project_descriptor)
-  - options = options_for_select(project_descriptors, @content.project_descriptor)
-  - if options.present?
-    = content.combo_field :project_descriptor, options, label: 'Project name',
-        required: true,
-        help: 'Please provide the <b>name of the project</b> the trial is part of. Either select an existing project name, or provide a new one.'
-  - else
-    = content.text_field :project_descriptor, required: true, label: 'Project name',
-      help: 'Please provide the <b>name of the project</b> the trial is part of.'
-
-.form-group
-  = content.label :plant_population_id, "Plant population", class: 'required'
-  - options = options_for_submission_select(@content.plant_population, :name)
-  = content.select :plant_population_id, options, { include_blank: true },
-    { class: 'form-control plant-population', data: { placeholder: 'Select plant population' } }
-  %small.help-block
-    If the trial was performed for a
-    %b population that is not listed in BIP,
-    please use Plant Population submission to upload information about that population first.
-
-.form-group
-  = content.text_field :trial_year, required: true,
-    placeholder: "Please provide information on the year the trial took place",
-    help: "Note that you can only submit trait data from <b>one year at a time</b>. Examples: 2004, 2005, 2016."
+.info-block
+  Before uploading your trait score sheet, please choose <b>existing trait descriptors</b>
+  that were scored in the course of this trial, or define <b>new trait descriptors.</b>
+  Further instructions are found below.
+  Examples of plant traits in the database are: 'oleic acid content' or ' seed mature time'.
 
 .form-group
   = label_tag "Submitted trait scoring data status"
@@ -48,27 +24,60 @@
     data), please do that using two separate Trial submissions.
 
 .form-group
-  = content.text_area :plant_trial_description, label: 'Trial description', required: true,
-    help: "Please provide a brief description of the trial. It should contain the trait(s) and the name of the population of interest, the trial type and location. The grant and project number could be named, too. An example of an existing trial description in the data base is: 'Seed quality assessment of the BnaTNDH population in a field trial in China'"
+  = content.label :trait_descriptor_list
+  - options = options_for_submission_select(@content.existing_trait_descriptors, :trait_name)
+  - options.concat options_for_submission_select(@content.new_trait_descriptors, :trait, id: false)
+  = content.select :trait_descriptor_list, options.html_safe, {},
+    { multiple: true,
+      class: 'trait-descriptor-list form-control',
+      'data-placeholder': "Select trait descriptors" }
 
-- %w(institute_id terrain soil_type statistical_factors).each do |attr|
-  .form-group
-    - values = PlantTrial.visible(current_user.id).attribute_values(attr)
-    - options = options_for_select(values, @content.send(attr))
-    = content.combo_field attr, options, required: (attr == 'institute_id')
+%button.btn.btn-default.new-trait-descriptor-for-list{ type: :button } Add new Trait descriptor
 
+.panel.panel-info.new-trait-descriptor-for-list.hidden
+  .panel-heading
+    %h4 Add new <b>Trait descriptor</b>
 
-%fieldset
-  %legend Location
+  .panel-body{"data-dirty-context": "new-trait-descriptor"}
+    .info-block
+      <b>Important:</b> If the trait you look for is missing you need to contact us
+      by #{support_email_link("email", "[BIP] New trait request")} and it will be added to our database.
 
-  .form-group= content.country_select :country_id, placeholder: 'Select a country', required: true
-  .form-group
-    = content.text_field :place_name, required: true,
-    help: 'Please provide a geographical name of the trial location, such as a town.'
-  .form-group
-    = content.text_field :trial_location_site_name,
-    help: "Please provide the name of the field, glasshouse or growth room used for the trial. Examples of place names in the database are: 'Wharf Ground' or 'unspecified glasshouse in Wellesbourne'."
-  .form-group= content.text_field :latitude, help: 'Expressed in degrees, eg. 48.86.'
-  .form-group= content.text_field :longitude, help: 'Expressed in degrees, eg. 2.30.'
-  .form-group= content.text_field :altitude, help: 'Expressed in meters.'
+    .form-group
+      = label_tag :trait, nil, class: 'required'
+      = select_tag :trait, [], class: 'trait form-control required', 'data-placeholder': 'Select a value'
+      %small.help-block Select the trait definition.
 
+    .form-group
+      - values = TraitDescriptor.visible(current_user.id).attribute_values('units_of_measurements')
+      - option_tags = options_for_select(values)
+      = combo_field_tag :units_of_measurements, nil, option_tags, class: 'units-of-measurements', required: true,
+          label: t('submission.steps.trial.step02.new_trait_descriptor_attributes.units_of_measurements')
+      %small.help-block
+        = t 'submission.steps.trial.step02.new_trait_descriptor_attribute_hints.units_of_measurements'
+
+    .form-group
+      = label_tag :scoring_method, nil, class: 'required'
+      = text_area_tag :scoring_method, nil, class: 'form-control required'
+      %small.help-block Describe in detail how the measurement is to be performed.
+
+    .form-group
+      = label_tag :materials
+      = text_area_tag :materials, nil, class: 'form-control'
+      %small.help-block Materials required for scoring.
+
+    .form-group
+      = label_tag :plant_part_id, nil
+      = select_tag :plant_part_id, [], class: 'plant-part-id form-control', 'data-placeholder': 'Select a plant part'
+      %small.help-block Select the part of plant one should use for scoring this trait.
+
+    = render partial: "submissions/annotations", locals: { model: TraitDescriptor }
+
+    .action-buttons
+      %button.btn.btn-default.cancel-new-trait-descriptor-for-list{type: 'button'} Cancel
+      %button.btn.btn-primary.add-new-trait-descriptor-for-list{type: 'button'} Confirm
+
+- @content.new_trait_descriptors.each do |trait_descriptor|
+  %div{id: "new-trait-descriptor-#{trait_descriptor.trait.split.join('-').downcase}"}
+    - @content.class.new_trait_descriptor_properties.each do |attr|
+      = hidden_field_tag "submission[content][new_trait_descriptors][][#{attr}]", trait_descriptor.send(attr)
