+++ app/views/submissions/steps/trial/_step02.html.haml
@@ -1,64 +1,88 @@
 .info-block
-  Before uploading your trait score sheet, please choose <b>existing trait descriptors</b>
-  that were scored in the course of this trial, or define <b>new trait descriptors.</b>
-  Further instructions are found below.
-  Examples of plant traits in the database are: 'oleic acid content' or ' seed mature time'.
+  %p
+    The BIP service assumes that each row in the trait scoring sheet, that you are going to submit,
+    represents a so-called <b>plant scoring unit</b>. For raw data it may simply be an individual
+    plant in your experiment. For processed data a plant scoring unit may represent a set of plants,
+    related to each other in some way. It is mandatory to provide plant accession information
+    for each plant scoring unit in the scoring sheet to be submitted.
 
-.form-group
-  = content.label :trait_descriptor_list
-  - options = options_for_submission_select(@content.existing_trait_descriptors, :trait_name)
-  - options.concat options_for_submission_select(@content.new_trait_descriptors, :trait, id: false)
-  = content.select :trait_descriptor_list, options.html_safe, {},
-    { multiple: true,
-      class: 'trait-descriptor-list form-control',
-      'data-placeholder': "Select trait descriptors" }
+  %p
+    In case you would like to check what plant accessions are currently registered in the BIP, please use the
+    = link_to 'plant accessions table view', data_tables_path(model: 'plant_accessions'), target: '_blank'
+    (you may filter data in that table with the <b>Search</b> text field).
 
-%button.btn.btn-default.new-trait-descriptor-for-list{ type: :button } Add new Trait descriptor
+  %p
+    If you submit raw data, you will have an option to submit technical replicate readings as
+    well as design factors (for instance, ordinal numbers for plots, reps and so forth).
 
-.panel.panel-info.new-trait-descriptor-for-list.hidden
-  .panel-heading
-    %h4 Add new <b>Trait descriptor</b>
+.fieldset
+  %legend Origin of genetic material used in this trial
+  .form-group
+    = content.label :lines_or_varieties, "Plant lines", value: 'plant_lines', class: 'radio-inline' do
+      = content.radio_button :lines_or_varieties, 'plant_lines', checked: true
+      Plant lines
+    = content.label :lines_or_varieties, "Plant varieties", value: 'plant_varieties', class: 'radio-inline' do
+      = content.radio_button :lines_or_varieties, 'plant_varieties'
+      Plant varieties
+    %small.help-block
+      Per each plant scoring unit submitted, the source of genetic material should be provided.
+      It may either be a <b>plant line</b> or a <b>plant variety (cultivar)</b>.
+      This choice depends on the type of experiment you are conducting. You should select
+      the <b>Plant lines</b> option if you
+      %ul
+        %li created new genetic material (using, for instance, crossing or mutagenesis) in the course of your project,
+        %li or you use former plant line genetic material which is not (yet) recognized as a variety or a cultivar.
+      In other cases you should probably select the the <b>Plant varieties</b> option.
 
-  .panel-body{"data-dirty-context": "new-trait-descriptor"}
+- if submission.raw_data?
+  %fieldset
+    %legend Design factors
     .info-block
-      <b>Important:</b> If the trait you look for is missing you need to contact us
-      by #{support_email_link("email", "[BIP] New trait request")} and it will be added to our database.
+      %p
+        Please select the appropriate statistical design factors suitable for your experimental setup from
+        the drop-down fields. Start by selecting the <b>highest design factor first</b>
+        and proceed with the next ones in <b>decreasing</b> size.
 
-    .form-group
-      = label_tag :trait, nil, class: 'required'
-      = select_tag :trait, [], class: 'trait form-control required', 'data-placeholder': 'Select a value'
-      %small.help-block Select the trait definition.
+      %p
+        If your factors are not present in the list provided, please add them manually. While creating new
+        factors, please make sure you comply with the rule to submit them in decreasing size, so with the
+        highest missing factor (e.g. treatment) first, to lowest missing factor (e.g. plot).
+        You are able to mix both predefined factors, from the drop-down list, and your own factor names.
 
-    .form-group
-      - values = TraitDescriptor.visible(current_user.id).attribute_values('units_of_measurements')
-      - option_tags = options_for_select(values)
-      = combo_field_tag :units_of_measurements, nil, option_tags, class: 'units-of-measurements', required: true,
-          label: t('submission.steps.trial.step02.new_trait_descriptor_attributes.units_of_measurements')
-      %small.help-block
-        = t 'submission.steps.trial.step02.new_trait_descriptor_attribute_hints.units_of_measurements'
+    - values = @content.design_factor_names
+    - (0...10).each do |idx|
+      .form-group{ class: (:hidden if idx > 0 && values[idx - 1].blank?) }
+        - options = options_for_select(DesignFactor::COMMON_NAMES, values[idx])
+        = content.combo_field "design_factor_names[]", options, value: values[idx],
+            label: "Design factor #{idx + 1}", idx: idx
 
-    .form-group
-      = label_tag :scoring_method, nil, class: 'required'
-      = text_area_tag :scoring_method, nil, class: 'form-control required'
-      %small.help-block Describe in detail how the measurement is to be performed.
+  %fieldset
+    %legend Technical replicate numbers
+    .info-block
+      %p
+        If you are using <i>technical replicates</i> in your experiment, please
+        set the number of replicates you are scoring per trait. Leaving the value
+        at <b>1</b> means you are not using technical replicates for a given trait.
 
-    .form-group
-      = label_tag :materials
-      = text_area_tag :materials, nil, class: 'form-control'
-      %small.help-block Materials required for scoring.
+      - if submission.sorted_trait_names.size > 1
+        %p
+          You can have varying number of technical replicates for different traits.
 
-    .form-group
-      = label_tag :plant_part_id, nil
-      = select_tag :plant_part_id, [], class: 'plant-part-id form-control', 'data-placeholder': 'Select a plant part'
-      %small.help-block Select the part of plant one should use for scoring this trait.
+    - sorted_trait_names = submission.sorted_trait_names
+    - col_count = [4, sorted_trait_names.size].min
+    - col_size = 12 / col_count
 
-    = render partial: "submissions/annotations", locals: { model: TraitDescriptor }
+    - sorted_trait_names.each_slice(col_count).with_index do |trait_names, row_idx|
+      .form-group
+        .row
+          - trait_names.each do |trait_name|
+            %div{ class: "col-md-#{col_size}" }= content.label :technical_replicate_numbers, trait_name
 
-    .action-buttons
-      %button.btn.btn-default.cancel-new-trait-descriptor-for-list{type: 'button'} Cancel
-      %button.btn.btn-primary.add-new-trait-descriptor-for-list{type: 'button'} Confirm
+        .row
+          - trait_names.each.with_index do |trait_name, col_idx|
+            - idx = row_idx * col_count + col_idx
 
-- @content.new_trait_descriptors.each do |trait_descriptor|
-  %div{id: "new-trait-descriptor-#{trait_descriptor.trait.split.join('-').downcase}"}
-    - @content.class.new_trait_descriptor_properties.each do |attr|
-      = hidden_field_tag "submission[content][new_trait_descriptors][][#{attr}]", trait_descriptor.send(attr)
+            %div{ class: "col-md-#{col_size}", data: { idx: idx } }
+              = content.text_field :technical_replicate_numbers,
+                  value: @content.technical_replicate_numbers[idx] || 1,
+                  label: false, name: "submission[content][technical_replicate_numbers][]"
