+++ app/views/submissions/steps/trial/_step08.html.haml
@@ -1,0 +1,64 @@
+.form-group
+  = content.text_field :plant_trial_name, required: true,
+    help: 'Please, provide convenient <b>name</b> for this scoring trial. It has to be <b>unique</b>.'
+
+.form-group
+  - project_descriptors = PlantTrial.visible(current_user.id).attribute_values(:project_descriptor)
+  - options = options_for_select(project_descriptors, @content.project_descriptor)
+  - if options.present?
+    = content.combo_field :project_descriptor, options, label: 'Project name',
+        required: true,
+        help: 'Please provide the <b>name of the project</b> the trial is part of. Either select an existing project name, or provide a new one.'
+  - else
+    = content.text_field :project_descriptor, required: true, label: 'Project name',
+      help: 'Please provide the <b>name of the project</b> the trial is part of.'
+
+.form-group
+  = content.label :plant_population_id, "Plant population", class: 'required'
+  - options = options_for_submission_select(@content.plant_population, :name)
+  = content.select :plant_population_id, options, { include_blank: true },
+    { class: 'form-control plant-population', data: { placeholder: 'Select plant population' } }
+  %small.help-block
+    If the trial was performed for a
+    %b population that is not listed in BIP,
+    please use Plant Population submission to upload information about that population first.
+
+.form-group
+  = content.text_field :trial_year, required: true,
+    placeholder: "Please provide information on the year the trial took place",
+    help: "Note that you can only submit trait data from <b>one year at a time</b>. Examples: 2004, 2005, 2016."
+
+.form-group
+  = content.text_area :plant_trial_description, label: 'Trial description', required: true,
+    help: "Please provide a brief description of the trial. It should contain the trait(s) and the name of the population of interest, the trial type and location. The grant and project number could be named, too. An example of an existing trial description in the data base is: 'Seed quality assessment of the BnaTNDH population in a field trial in China'"
+
+- %w(institute_id terrain soil_type statistical_factors).each do |attr|
+  .form-group
+    - values = PlantTrial.visible(current_user.id).attribute_values(attr)
+    - options = options_for_select(values, @content.send(attr))
+    = content.combo_field attr, options, required: (attr == 'institute_id')
+
+
+%fieldset
+  %legend Location
+
+  .form-group= content.country_select :country_id, placeholder: 'Select a country', required: true
+  .form-group
+    = content.text_field :place_name, required: true,
+    help: 'Please provide a geographical name of the trial location, such as a town.'
+  .form-group
+    = content.text_field :trial_location_site_name,
+    help: "Please provide the name of the field, glasshouse or growth room used for the trial. Examples of place names in the database are: 'Wharf Ground' or 'unspecified glasshouse in Wellesbourne'."
+  .form-group= content.text_field :latitude, help: 'Expressed in degrees, eg. 48.86.'
+  .form-group= content.text_field :longitude, help: 'Expressed in degrees, eg. 2.30.'
+  .form-group= content.text_field :altitude, help: 'Expressed in meters.'
+
+%fieldset
+  %legend Provenance
+
+  = render partial: "submissions/annotations", locals: { model: PlantTrial, collapsible: false, form: content }
+
+%fieldset
+  %legend Visibility status
+
+  = render partial: "submissions/steps/visibility", locals: { form: content }
