+++ spec/support/validation_matchers.rb
@@ -7,7 +7,7 @@ module ValidationMatchers
     validate_numericality_of(attr).is_greater_than_or_equal_to(-273.15)
   end
 
-  def validate_as_ratio(attr)
-    validate_numericality_of(attr).is_greater_than_or_equal_to(0).is_less_than_or_equal_to(1)
+  def validate_as_percentage(attr)
+    validate_numericality_of(attr).is_greater_than_or_equal_to(0).is_less_than_or_equal_to(100)
   end
 end
